refactor: extract class `DebuggingPreferencesDataStore.java`
Lines of code and McCabe Cyclomatic Complexity (CC) are reduced.
